+++ app/base.html
@@ -1,0 +1,49 @@
+<!doctype html>
+<html lang="">
+<head>
+    <meta charset="utf-8">
+    <!-- Setting IE=edge tells Internet Explorer to use the latest engine to render the page and execute Javascript -->
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="description" content="">
+    <!-- Mobile Safari introduced this tag to let web developers control the viewport's size and scale -->
+    <!-- The width property controls the size of the viewport, the initial-scale property controls the zoom level when the page is first loaded-->
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title></title>
+
+    <!--  Add to homescreen for Chrome on Android -->
+    <meta name="mobile-web-app-capable" content="yes">
+    <link rel="icon" sizes="192x192" href="/static/assets/images/chrome-ninja192-precomposed.png">
+
+    <!-- Add to homescreen for Safari on IOS -->
+    <meta name="apple-mobile-web-app-capable" content="yes">
+    <meta name="apple-mobile-web-app-status-bar-style" content="black">
+    <meta name="apple-mobile-web-app-title" content="React Webpack Node">
+    <link rel="apple-touch-icon-precomposed" href="/static/assets/images/apple-ninja152-precomposed.png">
+
+    <!-- Tile icon for Win8 (144x144 + tile color) -->
+    <meta name="msapplication-TileImage" content="/static/assets/images/ms-ninja144-precomposed.png">
+    <meta name="msapplication-TileColor" content="#3372DF">
+
+    <!-- SEO: If your mobile URL is different from the desktop URL, add a canonical link to the desktop page https://developers.google.com/webmasters/smartphone-sites/feature-phones -->
+    <!--
+    <link rel="canonical" href="http://www.example.com/">
+    -->
+    <link href='https://fonts.googleapis.com/css?family=Roboto+Condensed' rel='stylesheet' type='text/css'>
+    <link rel="stylesheet" href="/static/assets/styles/main.css">
+</head>
+<body>
+
+<div class="app">CONTENT</div>
+
+<!-- Google Analytics: change UA-XXXXX-X to be your site's ID -->
+<script>
+    (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
+        (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
+            m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
+    })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
+    ga('create', 'UA-XXXXX-X', 'auto');
+    ga('send', 'pageview');
+</script>
+<script async type="text/javascript" charset="utf-8" src="static/assets/app.js"></script>
+</body>
+</html>
